+++ TruePath.Tests/PathExtensionsTests.cs
@@ -1,56 +1,45 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+// SPDX-FileCopyrightText: 2024 TruePath contributors <https://github.com/ForNeVeR/TruePath>
+//
+// SPDX-License-Identifier: MIT
 
 namespace TruePath.Tests;
+
 public class PathExtensionsTests
 {
-    [Fact]
-    public void GetExtensionTests()
+    [Theory]
+    [InlineData("..", ".")]
+    [InlineData("foo/bar.txt", ".txt")]
+    [InlineData("/foo/bar.txt", ".txt")]
+    [InlineData("foo/bar.", ".")]
+    [InlineData("foo/bar", "")]
+    [InlineData(".gitignore", ".gitignore")]
+    public void GetExtensionWithDotTests(string path, string expected)
     {
-        IPath l = new LocalPath("foo/bar.txt");
-        IPath a = new AbsolutePath("/foo/bar.txt");
+        IPath local = new LocalPath(path);
+        Assert.Equal(expected,  local.GetExtensionWithDot());
 
-        Assert.Equal(".txt", l.GetExtension());
-        Assert.Equal(".txt", a.GetExtension());
+        if (!path.StartsWith('/')) return;
+
+        IPath a = new AbsolutePath(path);
+        Assert.Equal(expected, a.GetExtensionWithDot());
     }
 
-    [Fact]
-    public void GetExtension_OnInvalidFile_ThrowsArgumentException()
+    [Theory]
+    [InlineData(".", "")]
+    [InlineData("..", "")]
+    [InlineData("foo/bar.txt", "txt")]
+    [InlineData("/foo/bar.txt", "txt")]
+    [InlineData("foo/bar.", "")]
+    [InlineData("foo/bar", "")]
+    [InlineData(".gitignore", "gitignore")]
+    public void GetExtensionWithoutDotTests(string path, string expected)
     {
-        IPath a = new AbsolutePath("/foo/bar");
-        string expectedMessage = "bar is not a file";
-        var ex = Assert.Throws<ArgumentException>(() => a.GetExtension());
-        Assert.Equal(expectedMessage, ex.Message);
+        IPath l = new LocalPath(path);
+        Assert.Equal(expected, l.GetExtensionWithoutDot());
 
+        if (!path.StartsWith('/')) return;
 
-        IPath l = new LocalPath("foo/bar");
-        ex = Assert.Throws<ArgumentException>(() => l.GetExtension());
-        Assert.Equal(expectedMessage, ex.Message);
-    }
-
-    [Fact]
-    public void GetExtensionWithoutDotTests()
-    {
-        IPath l = new LocalPath("foo/bar.txt");
-        IPath a = new AbsolutePath("/foo/bar.txt");
-
-        Assert.Equal("txt", l.GetExtensionWithoutDot());
-        Assert.Equal("txt", a.GetExtensionWithoutDot());
-    }
-
-    [Fact]
-    public void GetExtensionWithoutDot_FileWithoutExtension_ThrowsArgumentException()
-    {
-        IPath a = new AbsolutePath("/foo/bar");
-        string expectedMessage = "bar is not a file";
-        var ex = Assert.Throws<ArgumentException>(() => a.GetExtensionWithoutDot());
-        Assert.Equal(expectedMessage, ex.Message);
-
-        IPath l = new LocalPath("foo/bar");
-        ex = Assert.Throws<ArgumentException>(() => l.GetExtensionWithoutDot());
-        Assert.Equal(expectedMessage, ex.Message);
+        IPath a = new AbsolutePath(path);
+        Assert.Equal(expected, a.GetExtensionWithoutDot());
     }
 }