+++ TruePath/LocalPath.cs
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2024 Friedrich von Never <friedrich@fornever.md>
+// SPDX-FileCopyrightText: 2024 TruePath contributors <https://github.com/ForNeVeR/TruePath>
 //
 // SPDX-License-Identifier: MIT
 
@@ -34,9 +34,6 @@ public readonly struct LocalPath(string value) : IEquatable<LocalPath>, IPath, I
 
     /// <inheritdoc cref="IPath.FileName"/>
     public string FileName => Path.GetFileName(Value);
-
-    /// <inheritdoc cref="IPath.FileEntryKind"/>
-    public FileEntryKind FileEntryKind => this.GetKind();
 
     /// <returns>The normalized path string contained in this object.</returns>
     public override string ToString() => Value;